% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{numeros_impares}
\alias{numeros_impares}
\title{Números impares de un vector}
\usage{
numeros_impares(x)
}
\arguments{
\item{x}{Un vector numérico}
}
\value{
Un vector con los números pares
}
\description{
Esta función toma un vector de números y devuelve un nuevo vector que
contiene solo los impares.
}
\details{
Si \code{x} no es un vector numérico, se produce un error.
}
\examples{
numeros_impares(c(-1, 4, 4.1, 0, 6))
numeros_impares(10:15)
numeros_impares(-2)

}
