% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{numeros_pares}
\alias{numeros_pares}
\title{Números pares de un vector}
\usage{
numeros_pares(x)
}
\arguments{
\item{x}{Un vector numérico}
}
\value{
Un vector con los números pares
}
\description{
Esta función toma un vector de números y devuelve un nuevo vector que
contiene solo los pares.
}
\details{
Si \code{x} no es un vector numérico, se produce un error.
}
\examples{
numeros_pares(c(-1, 4, 4.1, 0, 6))
numeros_pares(10:15)
numeros_pares(1)

}
